pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "omop-vocab-llm-rag"
version = "0.1.0"
description = "Map free-text lab events to OMOP/LOINC concepts via Claude + RAG."
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39",
    "sentence-transformers>=3.0",
    "torch>=2.2",
    "faiss-cpu>=1.8",
    "pandas>=2.1",
    "pyarrow>=15",
    "numpy>=1.26",
    "tqdm>=4.66",
]

[project.scripts]
omop-map = "omop_vocab_llm_rag.cli:main"

[tool.setuptools.packages.find]
include = ["omop_vocab_llm_rag*"]
